Block filter durability uses short
We were using byte before but a) bukkit/mc uses a short and b) byte was
breaking mod servers.

+++ src/main/java/me/botsko/prism/parameters/BlockParameter.java
@@ -33,7 +33,7 @@ public class BlockParameter extends SimplePrismParameterHandler {
                 if( b.contains( ":" ) && b.length() >= 3 ) {
                     final String[] ids = b.split( ":" );
                     if( ids.length == 2 && TypeUtils.isNumeric( ids[0] ) && TypeUtils.isNumeric( ids[1] ) ) {
-                        query.addBlockFilter( Integer.parseInt( ids[0] ), Byte.parseByte( ids[1] ) );
+                        query.addBlockFilter( Integer.parseInt( ids[0] ), Short.parseShort( ids[1] ) );
                     } else {
                         throw new IllegalArgumentException( "Invalid block name '" + b + "'. Try /pr ? for help" );
                     }
@@ -41,7 +41,7 @@ public class BlockParameter extends SimplePrismParameterHandler {
 
                     // It's id without a subid
                     if( TypeUtils.isNumeric( b ) ) {
-                        query.addBlockFilter( Integer.parseInt( b ), (byte) 0 );
+                        query.addBlockFilter( Integer.parseInt( b ), (short) 0 );
                     } else {
 
                         // Lookup the item name, get the ids
@@ -53,9 +53,9 @@ public class BlockParameter extends SimplePrismParameterHandler {
                                     // If we really care about the sub id
                                     // because it's a whole different item
                                     if( ItemUtils.dataValueUsedForSubitems( ids[0] ) ) {
-                                        query.addBlockFilter( ids[0], (byte) ids[1] );
+                                        query.addBlockFilter( ids[0], (short) ids[1] );
                                     } else {
-                                        query.addBlockFilter( ids[0], (byte) 0 );
+                                        query.addBlockFilter( ids[0], (short) 0 );
                                     }
                                 }
                             }
